requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests

